[DEV-6151] Update Legend color code
